gateway/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0.0
redis>=5.0.0
aiohttp>=3.9.0
pyjwt>=2.8.0
